+++ crates/hashi/src/communication/mod.rs
@@ -7,7 +7,9 @@
 #[cfg(test)]
 pub mod in_memory;
 pub mod interfaces;
+pub mod timeout_and_retry;
 
 #[cfg(test)]
 pub use in_memory::InMemoryOrderedBroadcastChannel;
 pub use interfaces::{ChannelError, ChannelResult, OrderedBroadcastChannel, P2PChannel};
+pub use timeout_and_retry::with_timeout_and_retry;
